chore: trimming js tests & improved workflows (#267)
