+++ .github/actions/benchmark/action.yml
@@ -1,0 +1,92 @@
+name: 'Run Benchmarks'
+description: 'Run Aztec benchmarks, generate diff report, and upload results'
+
+inputs:
+  pr-number:
+    description: 'Pull request number'
+    required: true
+  base-ref:
+    description: 'Base branch ref for comparison'
+    required: true
+  head-ref:
+    description: 'Head branch ref (source branch of PR)'
+    required: true
+  bench-dir:
+    description: 'Directory for benchmark files'
+    required: false
+    default: './benchmarks'
+
+runs:
+  using: 'composite'
+  steps:
+    - name: Encode base branch name
+      id: encode-base
+      shell: bash
+      run: |
+        # Replace slashes with hyphens for artifact naming
+        # GitHub artifact names don't support slashes
+        encoded=$(echo "${{ inputs.base-ref }}" | sed 's/\//-/g')
+        echo "base-ref-encoded=${encoded}" >> $GITHUB_OUTPUT
+
+        # Determine which workflow created the baseline
+        # dev/main baselines are from update-baseline.yml, others from pr-checks.yml
+        if [[ "${{ inputs.base-ref }}" == "dev" || "${{ inputs.base-ref }}" == "main" ]]; then
+          echo "workflow=update-baseline.yml" >> $GITHUB_OUTPUT
+        else
+          echo "workflow=pr-checks.yml" >> $GITHUB_OUTPUT
+        fi
+
+    - name: Download baseline artifact
+      id: download-baseline
+      continue-on-error: true
+      uses: dawidd6/action-download-artifact@0bd50d53a6d7fb5cb921e607957e9cc12b4ce392 # v12
+      with:
+        workflow: ${{ steps.encode-base.outputs.workflow }}
+        branch: ${{ inputs.base-ref }}
+        name: benchmark-baseline-${{ steps.encode-base.outputs.base-ref-encoded }}
+        path: ${{ inputs.bench-dir }}
+        if_no_artifact_found: warn
+
+    - name: Generate Markdown diff
+      uses: defi-wonderland/aztec-benchmark/action@chore/v4
+      with:
+        base_suffix: '_latest'
+        current_suffix: '_new'
+
+    - name: Comment diff on PR
+      uses: peter-evans/create-or-update-comment@v4
+      with:
+        issue-number: ${{ inputs.pr-number }}
+        body-path: benchmark-comparison.md
+
+    - name: Prepare benchmark files for upload
+      shell: bash
+      run: |
+        # Remove old _latest files (from base branch fetch) before renaming
+        rm -f ${{ inputs.bench-dir }}/*_latest.benchmark.json
+        
+        # Rename _new files to _latest (the canonical stored format)
+        for file in ${{ inputs.bench-dir }}/*_new.benchmark.json; do
+          if [ -f "$file" ]; then
+            new_name="${file/_new.benchmark.json/_latest.benchmark.json}"
+            echo "Renaming $file -> $new_name"
+            mv "$file" "$new_name"
+          fi
+        done
+
+    - name: Encode head branch name
+      id: encode-head
+      shell: bash
+      run: |
+        # Replace slashes with hyphens for artifact naming
+        # GitHub artifact names don't support slashes
+        encoded=$(echo "${{ inputs.head-ref }}" | sed 's/\//-/g')
+        echo "head-ref-encoded=${encoded}" >> $GITHUB_OUTPUT
+
+    - name: Upload branch baseline
+      uses: actions/upload-artifact@v4
+      with:
+        name: benchmark-baseline-${{ steps.encode-head.outputs.head-ref-encoded }}
+        path: ${{ inputs.bench-dir }}/*_latest.benchmark.json
+        retention-days: 90
+        if-no-files-found: error
