+++ .github/actions/js-tests/action.yml
@@ -1,0 +1,10 @@
+name: 'Run JS Tests'
+description: 'Run JavaScript integration tests'
+
+runs:
+  using: 'composite'
+  steps:
+    - name: Run JS tests
+      shell: bash
+      run: |
+        script -e -c "BASE_PXE_URL=http://localhost NODE_NO_WARNINGS=1 yarn test:js"
