+++ .github/actions/setup-aztec/action.yml
@@ -1,0 +1,81 @@
+name: 'Setup Aztec Environment'
+description: 'Common setup for Aztec development: Node, Foundry, Aztec CLI, dependencies, and local network'
+
+inputs:
+  start-pxe:
+    description: 'Whether to start PXE node'
+    required: false
+    default: 'false'
+  run-codegen:
+    description: 'Whether to run aztec codegen'
+    required: false
+    default: 'false'
+
+outputs:
+  aztec-version:
+    description: 'The detected Aztec version'
+    value: ${{ steps.aztec-version.outputs.version }}
+
+runs:
+  using: 'composite'
+  steps:
+    - name: Setup Node.js
+      uses: actions/setup-node@v4
+      with:
+        node-version: '22.17.0'
+
+    - name: Update PATH for Aztec
+      shell: bash
+      run: |
+        echo "$HOME/.aztec/current/bin" >> $GITHUB_PATH
+        echo "$HOME/.aztec/current/node_modules/.bin" >> $GITHUB_PATH
+        echo "$HOME/.aztec/bin" >> $GITHUB_PATH
+
+    - name: Detect Aztec version
+      id: aztec-version
+      shell: bash
+      run: |
+        VER=$(node -p "require('./package.json').config.aztecVersion")
+        echo "version=$VER" >> "$GITHUB_OUTPUT"
+        echo "Aztec version: $VER"
+
+    - name: Install Foundry
+      uses: foundry-rs/foundry-toolchain@v1
+
+    - name: Install Aztec CLI
+      shell: bash
+      run: |
+        INSTALL_URL="https://install.aztec.network/${{ steps.aztec-version.outputs.version }}/"
+        echo "Installing from: $INSTALL_URL"
+        curl -sL $INSTALL_URL > tmp.sh
+        VERSION=${{ steps.aztec-version.outputs.version }} bash tmp.sh <<< yes "yes"
+
+    - name: Switch CLI to target version
+      shell: bash
+      run: aztec-up install ${{ steps.aztec-version.outputs.version }}
+
+    - name: Start local network
+      if: inputs.start-pxe == 'true'
+      shell: bash
+      run: aztec start --local-network &
+
+    - name: Install dependencies
+      shell: bash
+      run: yarn --frozen-lockfile
+
+    - name: Compile contracts
+      shell: bash
+      run: aztec compile
+
+    - name: Codegen wrappers
+      if: inputs.run-codegen == 'true'
+      shell: bash
+      run: aztec codegen target --outdir src/artifacts -f
+
+    - name: Start PXE node
+      if: inputs.start-pxe == 'true'
+      shell: bash
+      run: |
+        VERSION=${{ steps.aztec-version.outputs.version }} aztec \
+          start --port 8081 --pxe --pxe.nodeUrl=http://localhost:8080/ \
+          --pxe.proverEnabled false &
