+++ .github/workflows/pr-checks.yml
@@ -1,0 +1,90 @@
+name: PR Checks
+
+on:
+  pull_request:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number }}
+  cancel-in-progress: true
+
+permissions:
+  pull-requests: write
+  issues: write
+
+env:
+  BENCH_DIR: ./benchmarks
+
+jobs:
+  # ══════════════════════════════════════════════════════════════════════════════
+  # BENCHMARK JOB
+  # ══════════════════════════════════════════════════════════════════════════════
+  benchmark:
+    name: Benchmark
+    runs-on: ubuntu-latest-m
+    timeout-minutes: 120
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          ref: ${{ github.event.pull_request.head.sha }}
+          fetch-depth: 0
+
+      - name: Setup Aztec environment
+        uses: ./.github/actions/setup-aztec
+        with:
+          start-pxe: 'true'
+          run-codegen: 'true'
+
+      - name: Run benchmarks
+        uses: ./.github/actions/benchmark
+        with:
+          pr-number: ${{ github.event.pull_request.number }}
+          base-ref: ${{ github.event.pull_request.base.ref }}
+          head-ref: ${{ github.event.pull_request.head.ref }}
+          bench-dir: ${{ env.BENCH_DIR }}
+
+  # ══════════════════════════════════════════════════════════════════════════════
+  # JS TESTS JOB
+  # ══════════════════════════════════════════════════════════════════════════════
+  js-tests:
+    name: JS Tests
+    runs-on: ubuntu-latest-m
+    timeout-minutes: 60
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Setup Aztec environment
+        uses: ./.github/actions/setup-aztec
+        with:
+          start-pxe: 'true'
+          run-codegen: 'true'
+
+      - name: Run JS tests
+        uses: ./.github/actions/js-tests
+
+  # ══════════════════════════════════════════════════════════════════════════════
+  # NOIR TESTS JOB
+  # ══════════════════════════════════════════════════════════════════════════════
+  noir-tests:
+    name: Noir Tests
+    runs-on: ubuntu-latest-m
+    timeout-minutes: 60
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Setup Aztec environment
+        uses: ./.github/actions/setup-aztec
+        with:
+          start-pxe: 'false'
+          run-codegen: 'false'
+
+      - name: Run Noir tests
+        uses: ./.github/actions/noir-tests
+        with:
+          threads: '12'
