+++ .github/workflows/release.yml
@@ -8,11 +8,19 @@ on:
         required: true
         type: string
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
+  # ══════════════════════════════════════════════════════════════════════════════
+  # PRODUCTION RELEASE JOB
+  # ══════════════════════════════════════════════════════════════════════════════
   release:
     name: Release
     environment: Production
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-latest-m
+    timeout-minutes: 60
     if: github.ref == 'refs/heads/main'
     permissions:
       contents: write  # Required to create and push tags
@@ -21,15 +29,17 @@ jobs:
       PROJECT_NAME: '@defi-wonderland/aztec-standards'
 
     steps:
-      - name: Checkout Repo
+      - name: Checkout
         uses: actions/checkout@v4
-
-      - name: Install Node
-        uses: actions/setup-node@v4
         with:
-          node-version: "22.17.0"
-          registry-url: "https://registry.npmjs.org"
-          cache: "yarn"
+          fetch-depth: 0
+
+      - name: Setup Aztec environment
+        id: setup-aztec
+        uses: ./.github/actions/setup-aztec
+        with:
+          start-pxe: 'false'
+          run-codegen: 'true'
 
       # TODO: Automate changing package.json version from the workflow input
       #       (e.g., open a PR or push a signed commit) so manual edits are not required.
@@ -43,41 +53,6 @@ jobs:
             echo "::error::Input version ($INPUT_VERSION) does not match package.json version ($PKG_VERSION). Update package.json or re-run with the matching version (no 'v' prefix)."
             exit 1
           fi
-
-      - name: Set up Docker
-        uses: docker/setup-buildx-action@v2
-
-      - name: Detect Aztec version
-        id: aztec-version
-        run: |
-          AZTEC_VERSION=$(node -p "require('./package.json').config.aztecVersion")
-          echo "version=$AZTEC_VERSION" >> "$GITHUB_OUTPUT"
-          echo "Aztec version is $version"
-
-      - name: Install Aztec CLI
-        run: |
-          curl -s https://install.aztec.network > tmp.sh
-          VERSION=${{ steps.aztec-version.outputs.version }} bash tmp.sh <<< yes "yes"
-      - name: Update path
-        run: echo "/home/runner/.aztec/bin" >> $GITHUB_PATH
-
-      - name: Set Aztec version
-        run: |
-          VERSION=${{ steps.aztec-version.outputs.version }} aztec-up
-
-      # This is a temporary hack to fix a problem with v3 releases.
-      - name: Manually tag the aztec version as `latest`
-        run: |
-          docker tag aztecprotocol/aztec:${{ steps.aztec-version.outputs.version }} aztecprotocol/aztec:latest
-
-      - name: Install dependencies
-        run: yarn --frozen-lockfile
-
-      - name: Compile
-        run: yarn compile
-
-      - name: Codegen
-        run: aztec codegen target --outdir artifacts
 
       - name: Compile artifacts to JS
         run: |
