--- .github/workflows/tests-js.yaml
@@ -1,85 +1,0 @@
-name: JS Tests
-
-on:
-    push:
-      branches:
-        - main
-    pull_request:
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref_name }}
-  cancel-in-progress: true
-
-jobs:
-    setup-and-run:
-      strategy:
-        matrix:
-          os: [ubuntu-latest-m]
-          threads: [12]
-
-      runs-on: 
-        labels: ${{ matrix.os }}
-      timeout-minutes: 60
-
-      concurrency:
-        group: job-${{ github.workflow }}-${{ github.ref_name }}-${{ matrix.os }}-${{ matrix.threads }}
-        cancel-in-progress: false
-
-      steps:
-        - name: Checkout repository
-          uses: actions/checkout@v3
-          with:
-            fetch-depth: 0 # Fetch all history for comparison
-  
-        - uses: actions/setup-node@v4
-          with:
-            node-version: "22.17.0"
-  
-        - name: Set up Docker
-          uses: docker/setup-buildx-action@v2
-  
-        - name: Detect Aztec version
-          id: aztec-version
-          run: |
-            AZTEC_VERSION=$(node -p "require('./package.json').config.aztecVersion")
-            echo "version=$AZTEC_VERSION" >> "$GITHUB_OUTPUT"
-            echo "Aztec version is ${{ steps.aztec-version.outputs.version }}"
-
-        - name: Install Foundry
-          uses: foundry-rs/foundry-toolchain@v1
-
-        - name: Install Aztec CLI
-          run: |
-            INSTALL_URL="https://install.aztec.network/${{ steps.aztec-version.outputs.version }}/"
-            echo "Installing from: $INSTALL_URL"
-            curl -sL $INSTALL_URL > tmp.sh
-            VERSION=${{ steps.aztec-version.outputs.version }} bash tmp.sh <<< yes "yes"
-
-        - name: Update path
-          run: |
-            echo "$HOME/.aztec/current/bin" >> $GITHUB_PATH
-            echo "$HOME/.aztec/current/node_modules/.bin" >> $GITHUB_PATH
-            echo "$HOME/.aztec/bin" >> $GITHUB_PATH
-  
-        - name: Set Aztec version
-          run: aztec-up install ${{ steps.aztec-version.outputs.version }}
-
-        - name: Start local-network
-          run: |
-            aztec start --local-network &
-
-        - name: Install project dependencies
-          run: yarn --frozen-lockfile
-  
-        - name: Compile
-          run: yarn compile
-  
-        - name: Codegen
-          run: yarn codegen
-
-        - name: Start PXE
-          run: |
-            aztec start --port 8081 --pxe --pxe.nodeUrl=http://localhost:8080/ --pxe.proverEnabled false &
-
-        - name: Run js tests
-          run: script -e -c "BASE_PXE_URL=http://localhost NODE_NO_WARNINGS=1 node --experimental-vm-modules $(yarn bin jest) --no-cache --runInBand --config jest.integration.config.json"
