+++ benchmarks/token_contract.benchmark.ts
@@ -8,7 +8,7 @@ import { parseUnits } from 'viem';
 // Import the new Benchmark base class and context
 import { Benchmark, BenchmarkContext } from '@defi-wonderland/aztec-benchmark';
 
-import { TokenContract } from '../artifacts/Token.js';
+import { TokenContract } from '../src/artifacts/Token.js';
 import { deployTokenWithMinter, initializeTransferCommitment, setupTestSuite } from '../src/ts/test/utils.js';
 
 // Extend the BenchmarkContext from the new package
