+++ benchmarks/tokenized_vault_contract.benchmark.ts
@@ -9,7 +9,7 @@ import { parseUnits } from 'viem';
 // Import the new Benchmark base class and context
 import { Benchmark, BenchmarkContext } from '@defi-wonderland/aztec-benchmark';
 
-import { TokenContract } from '../artifacts/Token.js';
+import { TokenContract } from '../src/artifacts/Token.js';
 import {
   deployVaultAndAssetWithMinter,
   setPrivateAuthWit,
