--- jest.integration.config.json
@@ -1,12 +1,0 @@
-{
-  "preset": "ts-jest/presets/default-esm",
-  "transform": {
-    "^.+\\.tsx?$": ["ts-jest", { "useESM": true }]
-  },
-  "moduleNameMapper": {
-    "^(\\.{1,2}/.*)\\.js$": "$1"
-  },
-  "testRegex": "./src/.*\\.test\\.ts$",
-  "rootDir": "./",
-  "testTimeout": 30000
-}
