+++ package.json
@@ -10,12 +10,11 @@
   "type": "module",
   "scripts": {
     "postinstall": "husky",
-    "clean": "rm -rf ./artifacts ./target codegenCache.json",
-    "codegen": "aztec codegen target --outdir artifacts --force",
+    "clean": "rm -rf ./src/artifacts ./target codegenCache.json",
+    "codegen": "aztec codegen target --outdir src/artifacts -f",
     "compile": "aztec compile",
-    "start:pxe": "docker compose -p sandbox -f ~/.aztec/docker-compose.sandbox.yml -f docker-compose.override.yml up",
     "test": "yarn test:nr && yarn test:js",
-    "test:js": "NODE_NO_WARNINGS=1 node --experimental-vm-modules $(yarn bin jest) --no-cache --runInBand --config jest.integration.config.json",
+    "test:js": "vitest run",
     "test:nr": "aztec test",
     "lint:prettier": "prettier '**/*.{js,ts}' --write",
     "bench": "NODE_NO_WARNINGS=1 aztec-benchmark --suffix _base",
@@ -33,26 +32,18 @@
     "@aztec/pxe": "4.0.0-nightly.20260204",
     "@aztec/stdlib": "4.0.0-nightly.20260204",
     "@aztec/test-wallet": "4.0.0-nightly.20260204",
-    "@defi-wonderland/aztec-benchmark": "github:defi-wonderland/aztec-benchmark#chore/v4",
-    "@types/node": "22.5.1"
+    "@defi-wonderland/aztec-benchmark": "github:defi-wonderland/aztec-benchmark#chore/v4"
   },
   "devDependencies": {
-    "@types/jest": "29.5.11",
-    "@types/mocha": "10.0.6",
-    "@types/node": "22.5.1",
+    "@commitlint/cli": "20.3.1",
+    "@commitlint/config-conventional": "20.2.0",
+    "@types/node": "25.0.10",
     "husky": "9.1.7",
-    "jest": "29.7.0",
-    "lint-staged": "15.4.3",
-    "prettier": "3.4.2",
-    "ts-jest": "29.2.5",
-    "ts-node": "10.9.2",
-    "typescript": "5.7.2"
-  },
-  "resolutions": {
-    "@aztec/foundation": "4.0.0-nightly.20260204"
-  },
-  "jest": {
-    "testTimeout": 200000
+    "lint-staged": "16.2.7",
+    "prettier": "3.8.0",
+    "tsx": "4.21.0",
+    "typescript": "5.9.3",
+    "vitest": "4.0.17"
   },
   "config": {
     "aztecVersion": "4.0.0-nightly.20260204"
