+++ src/ts/test/tokenized_vault.test.ts
@@ -11,7 +11,9 @@ import { type TestWallet } from '@aztec/test-wallet/server';
 import { type AztecLMDBStoreV2 } from '@aztec/kv-store/lmdb-v2';
 import { type ContractFunctionInteraction } from '@aztec/aztec.js/contracts';
 
-import { TokenContract } from '../../../artifacts/Token.js';
+import { describe, it, expect, beforeAll, beforeEach, afterAll } from 'vitest';
+
+import { TokenContract } from '../../../src/artifacts/Token.js';
 
 const TEST_TIMEOUT = 300_000;
 
@@ -19,7 +21,6 @@ describe('Tokenized Vault', () => {
   let store: AztecLMDBStoreV2;
   let wallet: TestWallet;
   let accounts: AztecAddress[];
-  let deployer: AztecAddress;
   let alice: AztecAddress;
   let bob: AztecAddress;
   let carl: AztecAddress;
@@ -710,572 +711,6 @@ describe('Tokenized Vault', () => {
         await expectTokenBalances(vault, bob, 0, 0);
         await expectTokenBalances(vault, carl, 0, 0);
         expect(await vault.methods.total_supply().simulate({ from: carl })).toBe(0n);
-      },
-      TEST_TIMEOUT,
-    );
-  });
-
-  describe('Deposit failures: incorrect amounts', () => {
-    // Alice exclusively interacts with the vault in this tests
-
-    it(
-      'deposit_public_to_public',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_public(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let transfer = asset.methods.transfer_public_to_public(alice, vault.address, initialAmount + 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.deposit_public_to_public(alice, alice, initialAmount + 1, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to deposit with an incorrect allowance
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, assetsAlice - 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.deposit_public_to_public(alice, alice, assetsAlice, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'deposit_public_to_private',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_public(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let transfer = asset.methods.transfer_public_to_public(alice, vault.address, initialAmount + 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_public_to_private(alice, alice, initialAmount + 1, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, assetsAlice - 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.deposit_public_to_private(alice, alice, assetsAlice, sharesRequested, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, assetsAlice, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.deposit_public_to_private(alice, alice, assetsAlice, sharesRequested, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Too many shares requested
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'deposit_private_to_public',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_private(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let transfer = asset.methods.transfer_private_to_public(alice, vault.address, initialAmount + 1, 0);
-        let transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_public(alice, alice, initialAmount + 1, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attemp to deposit with an incorrect allowance
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, assetsAlice - 1, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_public(alice, alice, assetsAlice, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Unknown auth witness for message hash /);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'deposit_private_to_private',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_private(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let transfer = asset.methods.transfer_private_to_public(alice, vault.address, initialAmount + 1, 0);
-        let transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private(alice, alice, initialAmount + 1, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, assetsAlice - 1, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private(alice, alice, assetsAlice, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Unknown auth witness for message hash /);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, assetsAlice, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private(alice, alice, assetsAlice, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Too many shares requested
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'deposit_public_to_private_exact',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_public(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let transfer = asset.methods.transfer_public_to_public(alice, vault.address, initialAmount + 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_public_to_private_exact(alice, alice, initialAmount + 1, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, assetsAlice - 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_public_to_private_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, assetsAlice, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_public_to_private_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'deposit_private_to_private_exact',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_private(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let transfer = asset.methods.transfer_private_to_public(alice, vault.address, initialAmount + 1, 0);
-        let transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private_exact(alice, alice, initialAmount + 1, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, assetsAlice - 1, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Unknown auth witness for message hash /);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, assetsAlice, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .deposit_private_to_private_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-  });
-
-  describe('Issue failures: incorrect amounts', () => {
-    // Alice exclusively interacts with the vault in this tests
-
-    it(
-      'issue_public_to_public',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_public(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let maxAssets = initialAmount + 1;
-        let transfer = asset.methods.transfer_public_to_public(alice, vault.address, maxAssets, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.issue_public_to_public(alice, alice, sharesRequested, maxAssets, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, maxAssets - 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.issue_public_to_public(alice, alice, sharesRequested, maxAssets, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'issue_public_to_private',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_public(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let maxAssets = initialAmount + 1;
-        let transfer = asset.methods.transfer_public_to_public(alice, vault.address, maxAssets, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.issue_public_to_private(alice, alice, sharesRequested, maxAssets, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, maxAssets - 1, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.issue_public_to_private(alice, alice, sharesRequested, maxAssets, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_public_to_public(alice, vault.address, maxAssets, 0);
-        await setPublicAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods.issue_public_to_private(alice, alice, sharesRequested, maxAssets, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'issue_private_to_public_exact',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_private(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let maxAssets = initialAmount + 1;
-        let transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets, 0);
-        let transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_public_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets - 1, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_public_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Unknown auth witness for message hash /);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_public_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'issue_private_to_private_exact',
-      async () => {
-        // Mint some assets to Alice
-        await asset.methods.mint_to_private(alice, initialAmount).send({ from: alice });
-
-        // Attempt depositing more assets than Alice actually has
-        let sharesRequested = initialAmount + 1;
-        let maxAssets = initialAmount + 1;
-        let transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets, 0);
-        let transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_private_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attemp to deposit with an incorrect allowance
-        sharesRequested = assetsAlice;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets - 1, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_private_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/Unknown auth witness for message hash /);
-
-        // Attemp to request more shares than allowed for the given deposit
-        sharesRequested = assetsAlice + 1;
-        maxAssets = assetsAlice;
-        transfer = asset.methods.transfer_private_to_public(alice, vault.address, maxAssets, 0);
-        transferAuthWitness = await setPrivateAuthWit(vault.address, transfer, alice, wallet);
-        await expect(
-          vault.methods
-            .issue_private_to_private_exact(alice, alice, sharesRequested, maxAssets, 0)
-            .with({ authWitnesses: [transferAuthWitness] })
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-  });
-
-  describe('Withdraw failures: incorrect amounts', () => {
-    // Alice exclusively interacts with the vault in this tests
-
-    it(
-      'withdraw_public_to_public',
-      async () => {
-        // Mint some assets to Alice in public and deposit to public shares.
-        await mintAndDepositInPublic(alice, initialAmount, assetsAlice);
-
-        // Attempt withdrawing more assets than allowed
-        await expect(
-          vault.methods.withdraw_public_to_public(alice, alice, assetsAlice + 1, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'withdraw_public_to_private',
-      async () => {
-        // Mint some assets to Alice in public and deposit to public shares.
-        await mintAndDepositInPublic(alice, initialAmount, assetsAlice);
-
-        // Attempt withdrawing more assets than allowed
-        await expect(
-          vault.methods.withdraw_public_to_private(alice, alice, assetsAlice + 1, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'withdraw_private_to_private',
-      async () => {
-        // Mint some assets to Alice in private and deposit to private shares.
-        await mintAndDepositInPrivate(alice, initialAmount, assetsAlice, sharesAlice);
-
-        // Attempt withdrawing more assets than allowed
-        let sharesRequested = assetsAlice;
-        await expect(
-          vault.methods
-            .withdraw_private_to_private(alice, alice, assetsAlice + 1, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attempt burning more shares than Alice actually has
-        sharesRequested = assetsAlice + 1;
-        await expect(
-          vault.methods
-            .withdraw_private_to_private(alice, alice, assetsAlice, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'withdraw_private_to_public_exact',
-      async () => {
-        // Mint some assets to Alice in private and deposit to private shares.
-        await mintAndDepositInPrivate(alice, initialAmount, assetsAlice, sharesAlice);
-
-        // Attempt withdrawing more assets than allowed
-        let sharesRequested = assetsAlice;
-        await expect(
-          vault.methods
-            .withdraw_private_to_public_exact(alice, alice, assetsAlice + 1, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // /Underflow/
-
-        // Attempt burning more shares than Alice actually has
-        sharesRequested = assetsAlice + 1;
-        await expect(
-          vault.methods
-            .withdraw_private_to_public_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'withdraw_private_to_private_exact',
-      async () => {
-        // Mint some assets to Alice in private and deposit to private shares.
-        await mintAndDepositInPrivate(alice, initialAmount, assetsAlice, sharesAlice);
-
-        // Attempt withdrawing more assets than allowed
-        let sharesRequested = assetsAlice;
-        await expect(
-          vault.methods
-            .withdraw_private_to_private_exact(alice, alice, assetsAlice + 1, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-
-        // Attempt burning more shares than Alice actually has
-        sharesRequested = assetsAlice + 1;
-        await expect(
-          vault.methods
-            .withdraw_private_to_private_exact(alice, alice, assetsAlice, sharesRequested, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-      },
-      TEST_TIMEOUT,
-    );
-  });
-
-  describe('Redeem failures: incorrect amounts', () => {
-    // Alice exclusively interacts with the vault in this tests
-
-    it(
-      'redeem_public_to_public',
-      async () => {
-        // Mint some assets to Alice in public and deposit to public shares.
-        await mintAndDepositInPublic(alice, initialAmount, assetsAlice);
-
-        // Attempt redeeming more shares than Alice actually has
-        let sharesRequested = assetsAlice + 1;
-        await expect(
-          vault.methods.redeem_public_to_public(alice, alice, sharesRequested, 0).send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // Underflow
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'redeem_public_to_private_exact',
-      async () => {
-        // Mint some assets to Alice in public and deposit to public shares.
-        await mintAndDepositInPublic(alice, initialAmount, assetsAlice);
-
-        // Attempt redeeming more shares than Alice actually has
-        let sharesRequested = assetsAlice + 1;
-        let minAssets = 1;
-        await expect(
-          vault.methods
-            .redeem_public_to_private_exact(alice, alice, sharesRequested, minAssets, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/); // /Underflow/
-
-        // Attempt redeeming with an invalid rate
-        sharesRequested = assetsAlice;
-        minAssets = assetsAlice + 1;
-        await expect(
-          vault.methods
-            .redeem_public_to_private_exact(alice, alice, sharesRequested, minAssets, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'redeem_private_to_public',
-      async () => {
-        // Mint some assets to Alice in private and deposit to private shares.
-        await mintAndDepositInPrivate(alice, initialAmount, assetsAlice, sharesAlice);
-
-        // Attempt redeeming more shares than Alice actually has
-        let sharesRequested = assetsAlice + 1;
-        await expect(
-          vault.methods.redeem_private_to_public(alice, alice, sharesRequested, 0).send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-      },
-      TEST_TIMEOUT,
-    );
-
-    it(
-      'redeem_private_to_private_exact',
-      async () => {
-        // Mint some assets to Alice in private and deposit to private shares.
-        await mintAndDepositInPrivate(alice, initialAmount, assetsAlice, sharesAlice);
-
-        // Attempt redeeming more shares than Alice actually has
-        let sharesRequested = assetsAlice + 1;
-        let minAssets = 1;
-        await expect(
-          vault.methods
-            .redeem_private_to_private_exact(alice, alice, sharesRequested, minAssets, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/Assertion failed: Balance too low 'subtracted > 0'/);
-
-        // Attempt redeeming with an invalid rate
-        sharesRequested = assetsAlice;
-        minAssets = assetsAlice + 1;
-        await expect(
-          vault.methods
-            .redeem_private_to_private_exact(alice, alice, sharesRequested, minAssets, 0)
-            .send({ from: alice }),
-        ).rejects.toThrow(/app_logic_reverted/);
       },
       TEST_TIMEOUT,
     );
